+++ dev/flipbook.tsx
@@ -1,4 +1,5 @@
 /* eslint monorepo/no-internal-import: "off", monorepo/no-relative-import: "off", import/no-relative-packages: "off" */
+import {scorePerseusItem} from "@khanacademy/perseus-score";
 import Banner from "@khanacademy/wonder-blocks-banner";
 import Button from "@khanacademy/wonder-blocks-button";
 import {View} from "@khanacademy/wonder-blocks-core";
@@ -18,7 +19,6 @@ import {useEffect, useMemo, useReducer, useRef, useState} from "react";
 
 import {Renderer} from "../packages/perseus/src";
 import {SvgImage} from "../packages/perseus/src/components";
-import {scorePerseusItem} from "../packages/perseus/src/renderer-util";
 import {mockStrings} from "../packages/perseus/src/strings";
 import {isCorrect} from "../packages/perseus/src/util/scoring";
 import {trueForAllMafsSupportedGraphTypes} from "../packages/perseus/src/widgets/interactive-graphs/mafs-supported-graph-types";
@@ -324,7 +324,6 @@ function GradableRenderer(props: QuestionRendererProps) {
                                 const score = scorePerseusItem(
                                     question,
                                     rendererRef.current.getUserInputMap(),
-                                    mockStrings,
                                     "en",
                                 );
                                 setScore(score);
