+++ packages/perseus-score/src/index.ts
@@ -36,3 +36,6 @@ export {
     default as scoreInputNumber,
     inputNumberAnswerTypes,
 } from "./widgets/input-number/score-input-number";
+
+export {scorePerseusItem, scoreWidgetsFunctional, flattenScores} from "./score";
+export {emptyWidgetsFunctional} from "./validate";
