+++ packages/perseus-score/src/score.test.ts
@@ -1,0 +1,470 @@
+import {
+    getExpressionWidget,
+    getLegacyExpressionWidget,
+    getTestDropdownWidget,
+} from "../util/test-helpers";
+
+import {flattenScores, scoreWidgetsFunctional} from "./score";
+
+import type {UserInputMap} from "./validation.types";
+import type {PerseusWidgetsMap} from "@khanacademy/perseus-core";
+
+describe("flattenScores", () => {
+    it("defaults to an empty score", () => {
+        const result = flattenScores({});
+
+        expect(result).toHaveBeenAnsweredCorrectly({shouldHavePoints: false});
+        expect(result).toEqual({
+            type: "points",
+            total: 0,
+            earned: 0,
+            message: null,
+        });
+    });
+
+    it("defaults to single score if there is only one", () => {
+        const result = flattenScores({
+            "radio 1": {
+                type: "points",
+                total: 1,
+                earned: 1,
+                message: null,
+            },
+        });
+
+        expect(result).toHaveBeenAnsweredCorrectly();
+        expect(result).toEqual({
+            type: "points",
+            total: 1,
+            earned: 1,
+            message: null,
+        });
+    });
+
+    it("returns an invalid score if any are invalid", () => {
+        const result = flattenScores({
+            "radio 1": {
+                type: "points",
+                total: 1,
+                earned: 1,
+                message: null,
+            },
+            "radio 2": {
+                type: "invalid",
+                message: null,
+            },
+        });
+
+        expect(result).toHaveInvalidInput();
+        expect(result).toEqual({
+            type: "invalid",
+            message: null,
+        });
+    });
+
+    it("tallies scores if multiple widgets have points", () => {
+        const result = flattenScores({
+            "radio 1": {
+                type: "points",
+                total: 1,
+                earned: 1,
+                message: null,
+            },
+            "radio 2": {
+                type: "points",
+                total: 1,
+                earned: 1,
+                message: null,
+            },
+            "radio 3": {
+                type: "points",
+                total: 1,
+                earned: 1,
+                message: null,
+            },
+        });
+
+        expect(result).toHaveBeenAnsweredCorrectly();
+        expect(result).toEqual({
+            type: "points",
+            total: 3,
+            earned: 3,
+            message: null,
+        });
+    });
+
+    it("doesn't count incorrect widgets", () => {
+        const result = flattenScores({
+            "radio 1": {
+                type: "points",
+                total: 1,
+                earned: 1,
+                message: null,
+            },
+            "radio 2": {
+                type: "points",
+                total: 1,
+                earned: 1,
+                message: null,
+            },
+            "radio 3": {
+                type: "points",
+                total: 1,
+                earned: 0,
+                message: null,
+            },
+        });
+
+        expect(result).toEqual({
+            type: "points",
+            total: 3,
+            earned: 2,
+            message: null,
+        });
+    });
+});
+
+describe("scoreWidgetsFunctional", () => {
+    it("returns an empty object when there's no widgets", () => {
+        // Arrange / Act
+        const result = scoreWidgetsFunctional({}, [], {}, "en");
+
+        // Assert
+        expect(result).toEqual({});
+    });
+
+    it("returns invalid if widget is unanswered", () => {
+        // Arrange
+        const widgets: PerseusWidgetsMap = {
+            "dropdown 1": getTestDropdownWidget(),
+        };
+        const widgetIds: Array<string> = ["dropdown 1"];
+        const userInputMap: UserInputMap = {
+            "dropdown 1": {
+                value: 0,
+            },
+        };
+
+        // Act
+        const result = scoreWidgetsFunctional(
+            widgets,
+            widgetIds,
+            userInputMap,
+
+            "en",
+        );
+
+        // Assert
+        expect(result["dropdown 1"]).toHaveInvalidInput();
+    });
+
+    it("can determine if a widget was answered correctly", () => {
+        // Arrange
+        const widgets: PerseusWidgetsMap = {
+            "dropdown 1": getTestDropdownWidget(),
+        };
+        const widgetIds: Array<string> = ["dropdown 1"];
+        const userInputMap: UserInputMap = {
+            "dropdown 1": {
+                value: 1,
+            },
+        };
+
+        // Act
+        const result = scoreWidgetsFunctional(
+            widgets,
+            widgetIds,
+            userInputMap,
+
+            "en",
+        );
+
+        // Assert
+        expect(result["dropdown 1"]).toHaveBeenAnsweredCorrectly();
+    });
+
+    it("can determine if a widget was answered incorrectly", () => {
+        // Arrange
+        const widgets: PerseusWidgetsMap = {
+            "dropdown 1": getTestDropdownWidget(),
+        };
+        const widgetIds: Array<string> = ["dropdown 1"];
+        const userInputMap: UserInputMap = {
+            "dropdown 1": {
+                value: 2,
+            },
+        };
+
+        // Act
+        const result = scoreWidgetsFunctional(
+            widgets,
+            widgetIds,
+            userInputMap,
+
+            "en",
+        );
+
+        // Assert
+        expect(result["dropdown 1"]).toHaveBeenAnsweredIncorrectly();
+    });
+
+    it("can handle multiple widgets", () => {
+        // Arrange
+        const widgets: PerseusWidgetsMap = {
+            "dropdown 1": getTestDropdownWidget(),
+            "dropdown 2": getTestDropdownWidget(),
+        };
+        const widgetIds: Array<string> = ["dropdown 1", "dropdown 2"];
+        const userInputMap: UserInputMap = {
+            "dropdown 1": {
+                value: 1,
+            },
+            "dropdown 2": {
+                value: 2,
+            },
+        };
+
+        // Act
+        const result = scoreWidgetsFunctional(
+            widgets,
+            widgetIds,
+            userInputMap,
+
+            "en",
+        );
+
+        // Assert
+        expect(result["dropdown 1"]).toHaveBeenAnsweredCorrectly();
+        expect(result["dropdown 2"]).toHaveBeenAnsweredIncorrectly();
+    });
+
+    it("skips widgets not in widgetIds", () => {
+        // Arrange
+        const widgets: PerseusWidgetsMap = {
+            "dropdown 1": getTestDropdownWidget(),
+            "dropdown 2": getTestDropdownWidget(),
+        };
+        const widgetIds: Array<string> = ["dropdown 1"];
+        const userInputMap: UserInputMap = {
+            "dropdown 1": {
+                value: 1,
+            },
+            "dropdown 2": {
+                value: 2,
+            },
+        };
+
+        // Act
+        const result = scoreWidgetsFunctional(
+            widgets,
+            widgetIds,
+            userInputMap,
+
+            "en",
+        );
+
+        // Assert
+        expect(result["dropdown 1"]).toHaveBeenAnsweredCorrectly();
+        expect(result["dropdown 2"]).toBe(undefined);
+    });
+
+    it("returns invalid if a widget in a group is unanswered", () => {
+        // Arrange
+        const widgets: PerseusWidgetsMap = {
+            "group 1": {
+                type: "group",
+                options: {
+                    content: "[[☃ dropdown 1]]",
+                    widgets: {
+                        "dropdown 1": getTestDropdownWidget(),
+                    },
+                    images: {},
+                },
+            },
+        };
+        const widgetIds: Array<string> = ["group 1"];
+        const userInputMap: UserInputMap = {
+            "group 1": {
+                "dropdown 1": {
+                    value: 0,
+                },
+            },
+        };
+
+        // Act
+        const result = scoreWidgetsFunctional(
+            widgets,
+            widgetIds,
+            userInputMap,
+
+            "en",
+        );
+
+        // Assert
+        expect(result["group 1"]).toHaveInvalidInput();
+    });
+
+    it("can score correct widget in group", () => {
+        // Arrange
+        const widgets: PerseusWidgetsMap = {
+            "group 1": {
+                type: "group",
+                options: {
+                    content: "[[☃ dropdown 1]]",
+                    widgets: {
+                        "dropdown 1": getTestDropdownWidget(),
+                    },
+                    images: {},
+                },
+            },
+        };
+        const widgetIds: Array<string> = ["group 1"];
+        const userInputMap: UserInputMap = {
+            "group 1": {
+                "dropdown 1": {
+                    value: 1,
+                },
+            },
+        };
+
+        // Act
+        const result = scoreWidgetsFunctional(
+            widgets,
+            widgetIds,
+            userInputMap,
+
+            "en",
+        );
+
+        // Assert
+        expect(result["group 1"]).toHaveBeenAnsweredCorrectly();
+    });
+
+    it("can score incorrect widget in group", () => {
+        // Arrange
+        const widgets: PerseusWidgetsMap = {
+            "group 1": {
+                type: "group",
+                options: {
+                    content: "[[☃ dropdown 1]]",
+                    widgets: {
+                        "dropdown 1": getTestDropdownWidget(),
+                    },
+                    images: {},
+                },
+            },
+        };
+        const widgetIds: Array<string> = ["group 1"];
+        const userInputMap: UserInputMap = {
+            "group 1": {
+                "dropdown 1": {
+                    value: 2,
+                },
+            },
+        };
+
+        // Act
+        const result = scoreWidgetsFunctional(
+            widgets,
+            widgetIds,
+            userInputMap,
+
+            "en",
+        );
+
+        // Assert
+        expect(result["group 1"]).toHaveBeenAnsweredIncorrectly();
+    });
+
+    it("can handle a correct modern Expression widget", () => {
+        // Arrange
+        const widgets: PerseusWidgetsMap = {
+            "expression 1": getExpressionWidget(),
+        };
+        const widgetIds: Array<string> = ["expression 1"];
+        const userInputMap: UserInputMap = {
+            "expression 1": "2+2",
+        };
+
+        // Act
+        const result = scoreWidgetsFunctional(
+            widgets,
+            widgetIds,
+            userInputMap,
+
+            "en",
+        );
+
+        // Assert
+        expect(result["expression 1"]).toHaveBeenAnsweredCorrectly();
+    });
+
+    it("can handle a correct legacy Expression widget", () => {
+        // Arrange
+        const widgets: PerseusWidgetsMap = {
+            "expression 1": getLegacyExpressionWidget() as any,
+        };
+        const widgetIds: Array<string> = ["expression 1"];
+        const userInputMap: UserInputMap = {
+            "expression 1": "2+2",
+        };
+
+        // Act
+        const result = scoreWidgetsFunctional(
+            widgets,
+            widgetIds,
+            userInputMap,
+
+            "en",
+        );
+
+        // Assert
+        expect(result["expression 1"]).toHaveBeenAnsweredCorrectly();
+    });
+
+    it("can handle an incorrect modern Expression widget", () => {
+        // Arrange
+        const widgets: PerseusWidgetsMap = {
+            "expression 1": getExpressionWidget(),
+        };
+        const widgetIds: Array<string> = ["expression 1"];
+        const userInputMap: UserInputMap = {
+            "expression 1": "2+42",
+        };
+
+        // Act
+        const result = scoreWidgetsFunctional(
+            widgets,
+            widgetIds,
+            userInputMap,
+
+            "en",
+        );
+
+        // Assert
+        expect(result["expression 1"]).toHaveBeenAnsweredIncorrectly();
+    });
+
+    it("can handle an incorrect legacy Expression widget", () => {
+        // Arrange
+        const widgets: PerseusWidgetsMap = {
+            "expression 1": getLegacyExpressionWidget() as any,
+        };
+        const widgetIds: Array<string> = ["expression 1"];
+        const userInputMap: UserInputMap = {
+            "expression 1": "2+42",
+        };
+
+        // Act
+        const result = scoreWidgetsFunctional(
+            widgets,
+            widgetIds,
+            userInputMap,
+
+            "en",
+        );
+
+        // Assert
+        expect(result["expression 1"]).toHaveBeenAnsweredIncorrectly();
+    });
+});
