+++ packages/perseus/src/index.ts
@@ -18,7 +18,6 @@ export {default as ServerItemRenderer} from "./server-item-renderer";
 export {default as HintsRenderer} from "./hints-renderer";
 export {default as HintRenderer} from "./hint-renderer";
 export {default as Renderer} from "./renderer";
-export {scorePerseusItem} from "./renderer-util";
 
 /**
  * Widgets
@@ -183,7 +182,6 @@ export {default as WIDGET_PROP_DENYLIST} from "./mixins/widget-prop-denylist";
 export type {ILogger, LogErrorOptions} from "./logging/log";
 export type {ServerItemRenderer as ServerItemRendererComponent} from "./server-item-renderer";
 export type {
-    Alignment,
     APIOptions,
     APIOptionsWithDefaults,
     ChangeHandler,
