+++ packages/perseus/src/renderer-util.test.ts
@@ -1,3 +1,4 @@
+import {scorePerseusItem} from "@khanacademy/perseus-score";
 import {act, screen} from "@testing-library/react";
 import {userEvent as userEventLib} from "@testing-library/user-event";
 
@@ -8,77 +9,11 @@ import {
 
 import {question1} from "./__testdata__/renderer.testdata";
 import * as Dependencies from "./dependencies";
-import {
-    emptyWidgetsFunctional,
-    scorePerseusItem,
-    scoreWidgetsFunctional,
-} from "./renderer-util";
-import {mockStrings} from "./strings";
 import {registerAllWidgetsForTesting} from "./util/register-all-widgets-for-testing";
 import {renderQuestion} from "./widgets/__testutils__/renderQuestion";
 import DropdownWidgetExport from "./widgets/dropdown";
 
-import type {
-    DropdownWidget,
-    ExpressionWidget,
-    PerseusWidgetsMap,
-} from "@khanacademy/perseus-core";
-import type {UserInputMap} from "@khanacademy/perseus-score";
 import type {UserEvent} from "@testing-library/user-event";
-
-function getTestDropdownWidget(): DropdownWidget {
-    return {
-        type: "dropdown",
-        options: {
-            choices: [
-                {
-                    content: "Test choice 1",
-                    correct: true,
-                },
-                {
-                    content: "Test choice 2",
-                    correct: false,
-                },
-            ],
-            placeholder: "Test placeholder",
-            static: false,
-        },
-    };
-}
-
-function getExpressionWidget(): ExpressionWidget {
-    return {
-        type: "expression",
-        version: {major: 1, minor: 0},
-        options: {
-            times: false,
-            buttonSets: ["basic"],
-            functions: [],
-            answerForms: [
-                {
-                    form: false,
-                    simplify: false,
-                    value: "2+2",
-                    considered: "correct",
-                },
-            ],
-        },
-    };
-}
-
-function getLegacyExpressionWidget() {
-    return {
-        type: "expression",
-        options: {
-            times: false,
-            buttonSets: ["basic"],
-            functions: [],
-            form: false,
-            simplify: false,
-            value: "2+2",
-        },
-    };
-}
 
 describe("renderer utils", () => {
     beforeAll(() => {
@@ -100,641 +35,6 @@ describe("renderer utils", () => {
                 ok: true,
             }),
         ) as jest.Mock;
-    });
-
-    describe("emptyWidgetsFunctional", () => {
-        it("returns an empty array if there are no widgets", () => {
-            // Arrange / Act
-            const result = emptyWidgetsFunctional(
-                {},
-                [],
-                {},
-                mockStrings,
-                "en",
-            );
-
-            // Assert
-            expect(result).toEqual([]);
-        });
-
-        it("properly identifies empty widgets", () => {
-            // Arrange
-            const widgets: PerseusWidgetsMap = {
-                "dropdown 1": getTestDropdownWidget(),
-            };
-            const widgetIds: Array<string> = ["dropdown 1"];
-            const userInputMap: UserInputMap = {
-                "dropdown 1": {
-                    value: 0,
-                },
-            };
-
-            // Act
-            const result = emptyWidgetsFunctional(
-                widgets,
-                widgetIds,
-                userInputMap,
-                mockStrings,
-                "en",
-            );
-
-            // Assert
-            expect(result).toEqual(["dropdown 1"]);
-        });
-
-        it("does not return widget IDs that are not empty", () => {
-            // Arrange
-            const widgets: PerseusWidgetsMap = {
-                "dropdown 1": getTestDropdownWidget(),
-            };
-            const widgetIds: Array<string> = ["dropdown 1"];
-            const userInputMap: UserInputMap = {
-                "dropdown 1": {
-                    value: 1,
-                },
-            };
-
-            // Act
-            const result = emptyWidgetsFunctional(
-                widgets,
-                widgetIds,
-                userInputMap,
-                mockStrings,
-                "en",
-            );
-
-            // Assert
-            expect(result).toEqual([]);
-        });
-
-        it("does not check for empty widgets whose IDs aren't provided", () => {
-            // Arrange
-            const widgets: PerseusWidgetsMap = {
-                "dropdown 1": getTestDropdownWidget(),
-            };
-            const widgetIds: Array<string> = [];
-            const userInputMap: UserInputMap = {
-                "dropdown 1": {
-                    value: 0,
-                },
-            };
-
-            // Act
-            const result = emptyWidgetsFunctional(
-                widgets,
-                widgetIds,
-                userInputMap,
-                mockStrings,
-                "en",
-            );
-
-            // Assert
-            expect(result).toEqual([]);
-        });
-
-        it("can properly split empty and non-empty widgets", () => {
-            // Arrange
-            const widgets: PerseusWidgetsMap = {
-                "dropdown 1": getTestDropdownWidget(),
-                "dropdown 2": getTestDropdownWidget(),
-            };
-            const widgetIds: Array<string> = ["dropdown 1", "dropdown 2"];
-            const userInputMap: UserInputMap = {
-                "dropdown 1": {
-                    value: 0,
-                },
-                "dropdown 2": {
-                    value: 1,
-                },
-            };
-
-            // Act
-            const result = emptyWidgetsFunctional(
-                widgets,
-                widgetIds,
-                userInputMap,
-                mockStrings,
-                "en",
-            );
-
-            // Assert
-            expect(result).toEqual(["dropdown 1"]);
-        });
-
-        it("properly identifies groups with empty widgets", () => {
-            // Arrange
-            const widgets: PerseusWidgetsMap = {
-                "group 1": {
-                    type: "group",
-                    options: {
-                        content: "[[☃ dropdown 1]]",
-                        widgets: {
-                            "dropdown 1": getTestDropdownWidget(),
-                        },
-                        images: {},
-                    },
-                },
-            };
-            const widgetIds: Array<string> = ["group 1"];
-            const userInputMap: UserInputMap = {
-                "group 1": {
-                    "dropdown 1": {
-                        value: 0,
-                    },
-                },
-            };
-
-            // Act
-            const result = emptyWidgetsFunctional(
-                widgets,
-                widgetIds,
-                userInputMap,
-                mockStrings,
-                "en",
-            );
-
-            // Assert
-            expect(result).toEqual(["group 1"]);
-        });
-
-        it("does not return group ID when its widgets are non-empty", () => {
-            // Arrange
-            const widgets: PerseusWidgetsMap = {
-                "group 1": {
-                    type: "group",
-                    options: {
-                        content: "[[☃ dropdown 1]]",
-                        widgets: {
-                            "dropdown 1": getTestDropdownWidget(),
-                        },
-                        images: {},
-                    },
-                },
-            };
-            const widgetIds: Array<string> = ["group 1"];
-            const userInputMap: UserInputMap = {
-                "group 1": {
-                    "dropdown 1": {
-                        value: 1,
-                    },
-                },
-            };
-
-            // Act
-            const result = emptyWidgetsFunctional(
-                widgets,
-                widgetIds,
-                userInputMap,
-                mockStrings,
-                "en",
-            );
-
-            // Assert
-            expect(result).toEqual([]);
-        });
-
-        it("handles an empty modern Expression widget", () => {
-            // Arrange
-            const widgets: PerseusWidgetsMap = {
-                "expression 1": getExpressionWidget(),
-            };
-            const widgetIds: Array<string> = ["expression 1"];
-            const userInputMap: UserInputMap = {
-                "expression 1": "",
-            };
-
-            // Act
-            const result = emptyWidgetsFunctional(
-                widgets,
-                widgetIds,
-                userInputMap,
-                mockStrings,
-                "en",
-            );
-
-            // Assert
-            expect(result).toEqual(["expression 1"]);
-        });
-
-        it("upgrades an empty legacy Expression widget", () => {
-            // Arrange
-            const widgets: PerseusWidgetsMap = {
-                "expression 1": getLegacyExpressionWidget() as any,
-            };
-            const widgetIds: Array<string> = ["expression 1"];
-            const userInputMap: UserInputMap = {
-                "expression 1": "",
-            };
-
-            // Act
-            const result = emptyWidgetsFunctional(
-                widgets,
-                widgetIds,
-                userInputMap,
-                mockStrings,
-                "en",
-            );
-
-            // Assert
-            expect(result).toEqual(["expression 1"]);
-        });
-
-        it("handles a non-empty modern Expression widget", () => {
-            // Arrange
-            const widgets: PerseusWidgetsMap = {
-                "expression 1": getExpressionWidget(),
-            };
-            const widgetIds: Array<string> = ["expression 1"];
-            const userInputMap: UserInputMap = {
-                "expression 1": "2+2",
-            };
-
-            // Act
-            const result = emptyWidgetsFunctional(
-                widgets,
-                widgetIds,
-                userInputMap,
-                mockStrings,
-                "en",
-            );
-
-            // Assert
-            expect(result).toEqual([]);
-        });
-
-        it("upgrades a non-empty legacy Expression widget", () => {
-            // Arrange
-            const widgets: PerseusWidgetsMap = {
-                "expression 1": getLegacyExpressionWidget() as any,
-            };
-            const widgetIds: Array<string> = ["expression 1"];
-            const userInputMap: UserInputMap = {
-                "expression 1": "2+2",
-            };
-
-            // Act
-            const result = emptyWidgetsFunctional(
-                widgets,
-                widgetIds,
-                userInputMap,
-                mockStrings,
-                "en",
-            );
-
-            // Assert
-            expect(result).toEqual([]);
-        });
-    });
-
-    describe("scoreWidgetsFunctional", () => {
-        it("returns an empty object when there's no widgets", () => {
-            // Arrange / Act
-            const result = scoreWidgetsFunctional(
-                {},
-                [],
-                {},
-                mockStrings,
-                "en",
-            );
-
-            // Assert
-            expect(result).toEqual({});
-        });
-
-        it("returns invalid if widget is unanswered", () => {
-            // Arrange
-            const widgets: PerseusWidgetsMap = {
-                "dropdown 1": getTestDropdownWidget(),
-            };
-            const widgetIds: Array<string> = ["dropdown 1"];
-            const userInputMap: UserInputMap = {
-                "dropdown 1": {
-                    value: 0,
-                },
-            };
-
-            // Act
-            const result = scoreWidgetsFunctional(
-                widgets,
-                widgetIds,
-                userInputMap,
-                mockStrings,
-                "en",
-            );
-
-            // Assert
-            expect(result["dropdown 1"]).toHaveInvalidInput();
-        });
-
-        it("can determine if a widget was answered correctly", () => {
-            // Arrange
-            const widgets: PerseusWidgetsMap = {
-                "dropdown 1": getTestDropdownWidget(),
-            };
-            const widgetIds: Array<string> = ["dropdown 1"];
-            const userInputMap: UserInputMap = {
-                "dropdown 1": {
-                    value: 1,
-                },
-            };
-
-            // Act
-            const result = scoreWidgetsFunctional(
-                widgets,
-                widgetIds,
-                userInputMap,
-                mockStrings,
-                "en",
-            );
-
-            // Assert
-            expect(result["dropdown 1"]).toHaveBeenAnsweredCorrectly();
-        });
-
-        it("can determine if a widget was answered incorrectly", () => {
-            // Arrange
-            const widgets: PerseusWidgetsMap = {
-                "dropdown 1": getTestDropdownWidget(),
-            };
-            const widgetIds: Array<string> = ["dropdown 1"];
-            const userInputMap: UserInputMap = {
-                "dropdown 1": {
-                    value: 2,
-                },
-            };
-
-            // Act
-            const result = scoreWidgetsFunctional(
-                widgets,
-                widgetIds,
-                userInputMap,
-                mockStrings,
-                "en",
-            );
-
-            // Assert
-            expect(result["dropdown 1"]).toHaveBeenAnsweredIncorrectly();
-        });
-
-        it("can handle multiple widgets", () => {
-            // Arrange
-            const widgets: PerseusWidgetsMap = {
-                "dropdown 1": getTestDropdownWidget(),
-                "dropdown 2": getTestDropdownWidget(),
-            };
-            const widgetIds: Array<string> = ["dropdown 1", "dropdown 2"];
-            const userInputMap: UserInputMap = {
-                "dropdown 1": {
-                    value: 1,
-                },
-                "dropdown 2": {
-                    value: 2,
-                },
-            };
-
-            // Act
-            const result = scoreWidgetsFunctional(
-                widgets,
-                widgetIds,
-                userInputMap,
-                mockStrings,
-                "en",
-            );
-
-            // Assert
-            expect(result["dropdown 1"]).toHaveBeenAnsweredCorrectly();
-            expect(result["dropdown 2"]).toHaveBeenAnsweredIncorrectly();
-        });
-
-        it("skips widgets not in widgetIds", () => {
-            // Arrange
-            const widgets: PerseusWidgetsMap = {
-                "dropdown 1": getTestDropdownWidget(),
-                "dropdown 2": getTestDropdownWidget(),
-            };
-            const widgetIds: Array<string> = ["dropdown 1"];
-            const userInputMap: UserInputMap = {
-                "dropdown 1": {
-                    value: 1,
-                },
-                "dropdown 2": {
-                    value: 2,
-                },
-            };
-
-            // Act
-            const result = scoreWidgetsFunctional(
-                widgets,
-                widgetIds,
-                userInputMap,
-                mockStrings,
-                "en",
-            );
-
-            // Assert
-            expect(result["dropdown 1"]).toHaveBeenAnsweredCorrectly();
-            expect(result["dropdown 2"]).toBe(undefined);
-        });
-
-        it("returns invalid if a widget in a group is unanswered", () => {
-            // Arrange
-            const widgets: PerseusWidgetsMap = {
-                "group 1": {
-                    type: "group",
-                    options: {
-                        content: "[[☃ dropdown 1]]",
-                        widgets: {
-                            "dropdown 1": getTestDropdownWidget(),
-                        },
-                        images: {},
-                    },
-                },
-            };
-            const widgetIds: Array<string> = ["group 1"];
-            const userInputMap: UserInputMap = {
-                "group 1": {
-                    "dropdown 1": {
-                        value: 0,
-                    },
-                },
-            };
-
-            // Act
-            const result = scoreWidgetsFunctional(
-                widgets,
-                widgetIds,
-                userInputMap,
-                mockStrings,
-                "en",
-            );
-
-            // Assert
-            expect(result["group 1"]).toHaveInvalidInput();
-        });
-
-        it("can score correct widget in group", () => {
-            // Arrange
-            const widgets: PerseusWidgetsMap = {
-                "group 1": {
-                    type: "group",
-                    options: {
-                        content: "[[☃ dropdown 1]]",
-                        widgets: {
-                            "dropdown 1": getTestDropdownWidget(),
-                        },
-                        images: {},
-                    },
-                },
-            };
-            const widgetIds: Array<string> = ["group 1"];
-            const userInputMap: UserInputMap = {
-                "group 1": {
-                    "dropdown 1": {
-                        value: 1,
-                    },
-                },
-            };
-
-            // Act
-            const result = scoreWidgetsFunctional(
-                widgets,
-                widgetIds,
-                userInputMap,
-                mockStrings,
-                "en",
-            );
-
-            // Assert
-            expect(result["group 1"]).toHaveBeenAnsweredCorrectly();
-        });
-
-        it("can score incorrect widget in group", () => {
-            // Arrange
-            const widgets: PerseusWidgetsMap = {
-                "group 1": {
-                    type: "group",
-                    options: {
-                        content: "[[☃ dropdown 1]]",
-                        widgets: {
-                            "dropdown 1": getTestDropdownWidget(),
-                        },
-                        images: {},
-                    },
-                },
-            };
-            const widgetIds: Array<string> = ["group 1"];
-            const userInputMap: UserInputMap = {
-                "group 1": {
-                    "dropdown 1": {
-                        value: 2,
-                    },
-                },
-            };
-
-            // Act
-            const result = scoreWidgetsFunctional(
-                widgets,
-                widgetIds,
-                userInputMap,
-                mockStrings,
-                "en",
-            );
-
-            // Assert
-            expect(result["group 1"]).toHaveBeenAnsweredIncorrectly();
-        });
-
-        it("can handle a correct modern Expression widget", () => {
-            // Arrange
-            const widgets: PerseusWidgetsMap = {
-                "expression 1": getExpressionWidget(),
-            };
-            const widgetIds: Array<string> = ["expression 1"];
-            const userInputMap: UserInputMap = {
-                "expression 1": "2+2",
-            };
-
-            // Act
-            const result = scoreWidgetsFunctional(
-                widgets,
-                widgetIds,
-                userInputMap,
-                mockStrings,
-                "en",
-            );
-
-            // Assert
-            expect(result["expression 1"]).toHaveBeenAnsweredCorrectly();
-        });
-
-        it("can handle a correct legacy Expression widget", () => {
-            // Arrange
-            const widgets: PerseusWidgetsMap = {
-                "expression 1": getLegacyExpressionWidget() as any,
-            };
-            const widgetIds: Array<string> = ["expression 1"];
-            const userInputMap: UserInputMap = {
-                "expression 1": "2+2",
-            };
-
-            // Act
-            const result = scoreWidgetsFunctional(
-                widgets,
-                widgetIds,
-                userInputMap,
-                mockStrings,
-                "en",
-            );
-
-            // Assert
-            expect(result["expression 1"]).toHaveBeenAnsweredCorrectly();
-        });
-
-        it("can handle an incorrect modern Expression widget", () => {
-            // Arrange
-            const widgets: PerseusWidgetsMap = {
-                "expression 1": getExpressionWidget(),
-            };
-            const widgetIds: Array<string> = ["expression 1"];
-            const userInputMap: UserInputMap = {
-                "expression 1": "2+42",
-            };
-
-            // Act
-            const result = scoreWidgetsFunctional(
-                widgets,
-                widgetIds,
-                userInputMap,
-                mockStrings,
-                "en",
-            );
-
-            // Assert
-            expect(result["expression 1"]).toHaveBeenAnsweredIncorrectly();
-        });
-
-        it("can handle an incorrect legacy Expression widget", () => {
-            // Arrange
-            const widgets: PerseusWidgetsMap = {
-                "expression 1": getLegacyExpressionWidget() as any,
-            };
-            const widgetIds: Array<string> = ["expression 1"];
-            const userInputMap: UserInputMap = {
-                "expression 1": "2+42",
-            };
-
-            // Act
-            const result = scoreWidgetsFunctional(
-                widgets,
-                widgetIds,
-                userInputMap,
-                mockStrings,
-                "en",
-            );
-
-            // Assert
-            expect(result["expression 1"]).toHaveBeenAnsweredIncorrectly();
-        });
     });
 
     describe("scorePerseusItem", () => {
@@ -773,7 +73,7 @@ describe("renderer utils", () => {
                     images: {},
                 },
                 {},
-                mockStrings,
+
                 "en",
             );
 
@@ -811,7 +111,7 @@ describe("renderer utils", () => {
                     images: {},
                 },
                 {"dropdown 1": {value: 0}},
-                mockStrings,
+
                 "en",
             );
 
@@ -840,7 +140,7 @@ describe("renderer utils", () => {
                     },
                 },
                 {"dropdown 1": {value: 1}},
-                mockStrings,
+
                 "en",
             );
 
@@ -865,7 +165,7 @@ describe("renderer utils", () => {
                     images: {},
                 },
                 {"dropdown 1": {value: 2}},
-                mockStrings,
+
                 "en",
             );
 
@@ -891,12 +191,7 @@ describe("renderer utils", () => {
 
             // Act
             const userInput = renderer.getUserInputMap();
-            const score = scorePerseusItem(
-                question1,
-                userInput,
-                mockStrings,
-                "en",
-            );
+            const score = scorePerseusItem(question1, userInput, "en");
 
             // Assert
             expect(score).toHaveBeenAnsweredCorrectly();
